Retrieve individual sighting data
