+++ src/App.js
@@ -1,30 +1,19 @@
 import React from "react";
 import logo from "./logo.png";
 import "./App.css";
-import { useState, useEffect } from "react";
-import axios from "axios";
 import List from "./Components/List";
+import Sighting from "./Components/Sighting";
+import { Routes, Route } from "react-router-dom";
 
 function App() {
-  const [sightings, setSightings] = useState([]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const { data } = await axios.get("http://localhost:3000/sightings");
-      console.log(data);
-      setSightings(data);
-    };
-
-    fetchData();
-  }, []);
-
   return (
     <div className="App">
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
-        {sightings.map((sighting, index) => (
-          <List sighting={sighting} index={index} />
-        ))}
+        <Routes>
+          <Route index element={<List />} />
+          <Route path="/sightings/:sightingsIndex" element={<Sighting />} />
+        </Routes>
       </header>
     </div>
   );
